refactored folders. added constants. updated click packages to the version 1.5.0
