+++ package.json
@@ -12,9 +12,9 @@
   },
   "dependencies": {
     "@apollo/client": "^3.10.8",
-    "@make-software/csprclick-core-client": "^1.4.5",
-    "@make-software/csprclick-core-types": "^1.4.5",
-    "@make-software/csprclick-ui": "^1.4.3",
+    "@make-software/csprclick-core-client": "^1.5.0",
+    "@make-software/csprclick-core-types": "^1.5.0",
+    "@make-software/csprclick-ui": "^1.5.0",
     "@metamask/safe-event-emitter": "^3.1.1",
     "casper-js-sdk": "^2.15.5",
     "classnames": "^2.5.1",
