+++ src/components/gettingStarted/components/Section.tsx
