+++ src/components/gettingStarted/components/Welcome.tsx
