+++ src/components/gettingStarted/components/index.tsx
