+++ src/components/gettingStarted/index.tsx
