+++ src/context/ContextComp.tsx
@@ -1,20 +1,23 @@
 'use client';
-import React, { ReactNode, useContext, useState } from 'react';
+import React, { ReactNode, useState } from 'react';
 
 export interface AppState {
   csprTheme: any;
   setCsprTheme: (csprTheme: any) => void;
 }
 
+const DARK_MODE = 'dark';
+
 const defaultState: AppState = {
-  csprTheme: 'dark',
-  setCsprTheme: () => {}
+  csprTheme: DARK_MODE,
+  setCsprTheme: () => {
+  }
 };
 
 const AppContext = React.createContext<AppState>(defaultState);
 
 const ContextComp = ({ children }: { children: ReactNode }) => {
-  const [csprTheme, setCsprTheme] = useState('dark');
+  const [csprTheme, setCsprTheme] = useState(DARK_MODE);
 
   return (
     <AppContext.Provider
@@ -27,7 +30,5 @@ const ContextComp = ({ children }: { children: ReactNode }) => {
     </AppContext.Provider>
   );
 };
-
-export const useAppContext = () => useContext(AppContext);
 
 export default ContextComp;
